// src/models/EquipamentoAeronave.ts
type EquipamentoAeronave = {
    NR_EQUIPAMENTO: number;
    NR_AERONAVE?: number;
    DS_CLASSE_PAT?: string;
    DS_SUB_ELEMENTO?: string;
    DS_TP_ABASTECIMENTO?: string;
    HR_TSN?: string;
    ID_PATRIMONIO?: number;
    NM_FANTASIA?: string;
    NR_MATRICULA?: number;
    NR_NSN?: string;
    NR_PATRIMONIO?: number;
    QT_AIFP?: string;
    QT_IPLR?: string;
    TP_ABASTECIMENTO?: string;
    VL_AQUISICAO?: number;
    VL_ATUALIZADO?: number;
    VL_DEPRECIACAO_MES?: number;
    VL_PERC_RESIDUAL?: number;
    VL_PRECO_EQP?: number;
    VL_RESIDUAL?: number;
    VL_VIDA_UTIL?: number;
    DT_VENC_ITEM?: string;
    VL_DISP_ITEM?: string;
    SG_INSPECAO_ITEM?: string;
    CD_UNIDADE?: string;
    CD_CODEMP?: string;
    CD_FORNECEDOR?: string;
    CD_MATERIAL?: number;
    CD_MATERIAL_CONFIG?: string;
    CD_MODELO?: string;
    CD_PROJETO?: string;
    CD_TABELA?: number;
    DS_CONFIGURACAO?: string;
    DS_SITUACAO_ATUAL?: string;
    DT_CONTRATO?: string;
    DT_FABRICACAO?: string;
    DT_FIM_INSPECAO?: string;
    DT_INICIO_DISPONIBILIDADE?: string;
    DT_RECEBIMENTO?: string;
    IN_ATIVO?: string;
    IN_DISPONIBILIDADE?: string;
    IN_SITUACAO_ATUAL?: string;
    NM_BASICO?: string;
    NM_TABELA?: string;
    NR_CONFIGURACAO?: number;
    NR_CONTRATO?: string;
    NR_PN?: string;
    NR_SERIE?: string;
    SG_MODELO?: string;
    SG_PROJETO?: string;
    SG_UNIDADE?: string;
    TP_EQUIPAMENTO?: string;
    TX_DISPONIBILIDADE?: string,
    TX_OBS?: string,
};

export default EquipamentoAeronave;