// src/models/ConfiguracaoPrimaria.ts
type ConfiguracaoPrimaria = {
    CD_MATERIAL: number;
    NR_CONFIGURACAO: number;
    CD_ITEM?: number;
    DS_CONFIGURACAO: string;
    DS_LOCAL?: string;
    DT_CONFIGURACAO?: string;
    DT_INSERT?: string;
    DT_UPDATE?: string;
    IN_ATIVO?: string;
    NM_USER_INSERT?: string;
    NM_USER_UPDATE?: string;
    QT_MATERIAL?: number;
}

export default ConfiguracaoPrimaria;